requirements.txt
adbc-driver-flightsql==1.8.0
adbc-driver-manager==1.8.0
agate==1.9.1
alembic==1.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.11.0
attrs==25.4.0
authlib==1.6.4
babel==2.17.0
backoff==2.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
coloredlogs==14.0
cryptography==46.0.2
daff==1.4.2
dagster==1.11.14
dagster-dbt==0.27.13
dagster-dlt==0.27.13
dagster-graphql==1.11.14
dagster-pipes==1.11.14
dagster-shared==1.11.14
dagster-webserver==1.11.14
dbt-adapters==1.17.2
dbt-common==1.32.0
dbt-core==1.10.13
dbt-duckdb==1.9.6
dbt-extractor==0.6.0
dbt-mcp==0.10.3
dbt-protos==1.0.382
dbt-semantic-interfaces==0.9.0
dbt-sl-sdk==0.13.0
dbtlabs-vortex==0.2.0
deepdiff==8.6.1
distro==1.9.0
dlt==1.17.1
docstring-parser==0.17.0
duckdb==1.4.1
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.9.0
gitdb==4.0.12
gitpython==3.1.45
giturlparse==0.12.0
gql==3.5.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
griffe==1.14.0
grpcio==1.75.1
grpcio-health-checking==1.75.1
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.2
humanfriendly==10.0
humanize==4.13.0
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
isodate==0.6.1
jinja2==3.1.6
jiter==0.11.0
jsonpath-ng==1.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
leather==0.4.0
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
mashumaro==3.14
mcp==1.17.0
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.1
multidict==6.7.0
networkx==3.5
numpy==2.3.3
openai==1.109.1
openai-agents==0.3.3
orderly-set==5.5.0
orjson==3.11.3
packaging==25.0
pandas==2.2.3
parsedatetime==2.6
pathlib-abc==0.5.1
pathspec==0.12.1
pathvalidate==3.3.1
pendulum==3.1.0
pluggy==1.6.0
ply==3.11
propcache==0.4.0
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydantic==2.11.10
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
requirements-parser==0.13.0
rich==14.1.0
rich-argparse==1.7.1
rpds-py==0.27.1
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowplow-tracker==1.1.0
sqlalchemy==2.0.43
sqlglot==27.20.0
sqlglotrs==0.6.2
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.47.3
structlog==25.4.0
tabulate==0.9.0
tenacity==9.1.2
text-unidecode==1.3
tomli==2.2.1
tomlkit==0.13.3
toposort==1.10
tqdm==4.67.1
typer==0.19.2
types-requests==2.32.4.20250913
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
universal-pathlib==0.3.2
urllib3==2.5.0
uuid==1.30
uvicorn==0.37.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
